Report an error early if the target exists and is not a plain file.
While we cannot prevent the target from being (re)created while the temporary
is being written, this check avoids an obvious eventual failure when the target
exists as a directory. Moreover, if the target exists and is some other type of
file (a link, a device, etc.) this reduces likelihood of an error of intent.